KUDU-2427: relax dlsym call to dl_iterate_phdr
When gcc passes --as-needed to ld by default, dynamically linked test
binaries that don't directly reference symbols in kudu_util may end up
loading libc before loading kudu_util. This causes dlsym(RTLD_NEXT, ...) on
libc symbols to fail because libc falls earlier in the symbol search order.

At first I thought this was a new problem in Ubuntu 18.04, but I was able to
reproduce it in Ubuntu 16.04 too; the only difference is that in Ubuntu
16.04 the failure isn't reported via dlerror() so we never noticed it. Given
that, I figured it'd be okay to make it non-fatal across the board.

Change-Id: I1795847740b1201c46331b6a64a96631ecbcea36
Reviewed-on: http://gerrit.cloudera.org:8080/10434
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Adar Dembo <adar@cloudera.com>