[Java] Add the RangeSchema.getColumnIds
Currently RangeSchema has a method that returns column
IDs called getColumns. Because the return type is a list
of integers it’s unclear if the integers are column indexes
or column IDs.

This change matches the HashBucketSchema by
naming the method getColumnIds. It also deprecates
the old method, and adds Javadoc to make the return
value clear.

Change-Id: Ie4fd944d6b38a2b3fd0ab7f8d6c254be380b2972
Reviewed-on: http://gerrit.cloudera.org:8080/10390
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Dan Burkert <danburkert@apache.org>