KUDU-2427: retry more system calls on EINTR
In order to collect its own stack traces, Kudu periodically sends itself a
SIGUSR2. The diagnostics log initiates stack collection every 60s, as do
some service queue overflow events. In theory, the collection shouldn't
affect any ongoing syscalls because the SIGUSR2 signal handler is installed
with SA_RESTART; in practice, not all syscalls are restartable, and
precisely categorizing those that are and those that aren't is difficult. As
such, it's really important that we retry every interruptible syscall rather
than surfacing the EINTR up the call stack as a failure.

For whatever reason this happens more frequently on Ubuntu 18.04, though
maybe it's because I've placed my test directory on tmpfs. For example, I
can easily repro a crash due to non-existent retry with the following
command line:

  bin/tablet_server-test --gtest_repeat=1000 --gtest_throw_on_failure \
    --diagnostics_log_stack_traces_interval_ms=100 \
    --unlock_experimental_flags --gtest_filter=*KUDU_177

This patch also fixes KUDU-2151.

Change-Id: I6cce03c4e1b2be32c1910382737526082fc99966
Reviewed-on: http://gerrit.cloudera.org:8080/10435
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>