[tools] Fix and simplify tablet summary spacing
ee0a75c5a2 attempted to fix spacing of tablet summaries but
messed up the case when a tablet is healthy, causing ksck to
print an empty line for every healthy tablet. This fixes that
and simplifies how the spacing is done.

I tested manually that this produces well-spaced output in the
three cases:
1. No tablets.
2. All healthy tablets in PLAIN_CONCISE mode (so no tablet info printed).
3. Some unhealthy tablets.

Change-Id: Ia1a9acfc27733e9c1df15cfe5cfc7822b0ca7304
Reviewed-on: http://gerrit.cloudera.org:8080/10416
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>