[java] Remove private method usage in kudu-mapreduce
kudu-mapreduce uses private APIs from KuduPredicate.
This is an issue because the API returns a protobuf
generated class and uses shaded APIs.

This isn’t strictly an issue in the gradle build itself because
calling the shaded methods is ok, but it is an issue in IDEs
like Intellij.

Regardless a shaded dependency is not part of
the public API and should not be exposed/used
because it can cause problems like this.

A follow on patch will work on enforcing that
protobuf is not exposed in the client API so this doesn’t
happen again.

Change-Id: I7ce54f715f970d1c7f9433e39af27d618e73fe6a
Reviewed-on: http://gerrit.cloudera.org:8080/10424
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>