KUDU-2444 TabletServerTest.TestTombstonedTabletOnWebUI is flaky
It's possible for a tablet to be RUNNING with working consensus
but still have its "creating" transition live in the tablet
manager. This makes deleting a tablet without retry immediately
after it is created flaky. To fix this for tests using an
in-process tablet server, I added a method to the TSTabletManager
that waits for there to be no tablet transitions in progress.
This method is intended for use only in tests, and is used by the
tablet server test fixture to ensure its single tablet has transitioned
out of the creating state before the actual test runs.

In addition to the original report from tablet_server-test.cc,
I also saw a test in tablet_copy_client_session-itest.cc fail
because of the same issue. I changed it to retry the delete, as
this tests uses an external minicluster.

I ran the affected tests 1000 times with 8 stress threads in
debug mode in dist-test. All passed.

Change-Id: Ib8dc7acc4c39f9d9b067a52df4bd6732f545a8de
Reviewed-on: http://gerrit.cloudera.org:8080/10443
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>