KUDU-2191: Leave table type as it is when populate a table
Now in HmsCatalog we always set table type to be managed when
populating a table. However, PopulateTable() is used in AlterTable as
well which implies the table type for existing tables in the Hive
Metastore may change without users' knowledge. Moreover,
PopulateTable() can be used in the metadata consistency check tool,
which may result in validation failure for external tables, though
there is none.

By default, when creating a table in the Hive Metastore, the table type
is managed without specifying the type. Thus, there is not much value
to specify it anyway.

Change-Id: I8cff392802308a5ea6262c17d248d37e00ac8190
Reviewed-on: http://gerrit.cloudera.org:8080/10477
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>