[tools] ksck imp. [8/n]: Easy-to-see errors when running w/o admin
If ksck is not run as an admin user, then it cannot gather all of the
information it uses to do cluster checks. "Not authorized" errors are
easy to mistake for other errors. This patch handles these errors
specially to make it obvious to administrators when they have forgotten
to run as an administrative user.

The new errors at the end of the ksck output look like this:

==================
Errors:
==================
Not authorized: error fetching info from masters: failed to gather info from 1 of 1 masters due to lack of admin privileges
Not authorized: error fetching info from tablet servers: failed to gather info from 11 of 12 tablet servers due to lack of admin privileges

FAILED
Not authorized: re-run ksck with administrator privileges

Change-Id: I644957103efefceeba4b82f4557376a603507423
Reviewed-on: http://gerrit.cloudera.org:8080/10300
Reviewed-by: Attila Bukor <abukor@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>