[tools] Add version info and check to ksck
This adds a new version summary table, similar to the one on the
master web ui for tablet servers, that summarizes the versions of Kudu
running in the cluster (both masters and tablet servers). ksck will
issue a warning if there are multiple versions. The version info is
available as part of the server health summaries in the KsckResults, so
clients of the Ksck class can use this information. Version info is
plumbed into the JSON output as well.

An example table and warning:

Version Summary
    Version     |                                                  Servers
----------------+-----------------------------------------------------------------------------------------------------
 1.6.0          | master@master-0.foo.com, tserver@tserver-0.foo.com, tserver@tserver-1.foo.com, and 1 other server(s)
 1.7.0          | master@master-1.foo.com
 1.8.0-SNAPSHOT | master@master-2.foo.com

==================
Warnings:
==================
version check error: not all servers are running the same version: 3 different versions were seen

Change-Id: I4a2bc7138f074ab8d74f21ace2eb2b8018c53f50
Reviewed-on: http://gerrit.cloudera.org:8080/10579
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>