KUDU-2191: Add a method to decide the scope of AlterTable
This commit adds a private method in KuduTableAlterer to indicate
whether the table alteration should be applied to external catalogs,
such as the Hive Metastore, which the Kudu master may has been
configured to integrate with.

This method will be useful to the HMS integration tools, such as the
metadata upgrade tool which may need to alter tables in Kudu only.

Change-Id: I703f12b95bb7ca2d65455f0f0602520332b3c678
Reviewed-on: http://gerrit.cloudera.org:8080/10581
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>