KUDU-1038 Deleting a tablet should also delete its log recovery directory, if any
This fix ensures successful bootstrap of a failed tablet replica, in the case
when its WAL segments are missing/deleted. The following scenario is tested:
1. A log segment is deleted from one of the Tablet Replicas.

2. On server restart, the replica in step 1. enters a failed state.
( Without this fix, due to the presence of log recovery directory (with deleted
log segments), the replica remains in a FAILED state and fails to bootstrap )

3. This fix deletes the log recovery directory when the tablet data is deleted,
as a result of which the failed replica successfully bootstraps.

Change-Id: If3f017caf41ac4e00bf89421d6e73924b9fdcbd0
Reviewed-on: http://gerrit.cloudera.org:8080/10591
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>