[rebalancer] 'auto' mode for RF=1 tablet movements
Added logic to automatically determine whether the tool should
move replicas of tablets with RF=1.  This is determined by examining the
version of Kudu and the replica management schema in the cluster, which
indicates whether the fix for KUDU-2443 applies.  The fix is present
whenever the version is larger than 1.7.0.

This patch also contains a unit test for the newly introduced
Is343SchemeCluster() utility function and an integration test scenario
RebalancerAndSingleReplicaTablets.  Also, already existing
RebalanceParamTest provides additional coverage for the functionality
in case of --move_single_replicas=auto (the default setting).

Change-Id: Id5a06b137cb34d9351f7e2fb819ed52790e2ee7e
Reviewed-on: http://gerrit.cloudera.org:8080/10612
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins