Increase --tablet_copy_begin_session_timeout_ms to 30 sec
This patch bumps the tablet copy begin session timeout from 3 seconds to
30 seconds. 3 seconds is too low of a value in cases where there is a
lot of disk activity on the remote server.

This patch also tweaks a couple of other minor things, like explicitly
setting the controller timeout in the case of a StartTabletCopy() RPC,
however there is no functionality change there because the same
RpcController is shared between UpdateConsensus() and StartTabletCopy()
calls, and the same timeout value is used for both. I just found it
confusing that the timeout was explicitly set for one but not the other.

Change-Id: I9e364d60351ad925583f8d7ef9dea55e0dac2507
Reviewed-on: http://gerrit.cloudera.org:8080/10651
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Tested-by: Kudu Jenkins