KUDU-2459: add placeholder names to some CREATE TABLE statements
Under the Tables section of Kudu Web UI, for a selected table, the table
metrics display a CREATE TABLE statement that can be run to make Impala
cognizant of that table. However, in generation of this statement, the
tablename tries to match the original Kudu tablename which may not
always be acceptable as a tablename for Impala. For example, Kudu
accepts dot in tablename, Impala does not. The CREATE TABLE statement
thus throws an invalid tablename error in Impala.

We considered trying to derive an Impala-conforming name from the Kudu-
supplied tablename but that could result in surprising/unintuitive
results. We also want to leverage the current functionality of auto-
generated statement where the name is valid. Therefore with this update,
invalid tablenames have been identified and updated by a placeholder
tablename for end user to replace with a valid tablename. The rules to
check Impala conformity are below --

- The minimum length of an identifier is 1 character.

- The maximum length of an identifier is currently 128 characters.

- An identifier must start with an alphabetic character.
The remainder can contain any combination of alphanumeric characters and
underscores. Quoting the identifier with backticks has no effect on the
allowed characters in the name.

- An identifier can contain only ASCII characters.

Change-Id: If9e5242318589452a0507d3973b9be3e6a186a69
Reviewed-on: http://gerrit.cloudera.org:8080/10656
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins