[tools] Add simple util for version parsing
This adds a struct and parsing function to parse versions of the form

<major>.<minor>.<maintenance>-<extra>

This can be used along with ksck's version reporting or ad hoc
version-gathering to determine how a tool should function when operating
against a cluster running certain Kudu versions.

Its first application will be in a follow-up.

Change-Id: I0bcf3b10b699e89ad489144611990702dec8af19
Reviewed-on: http://gerrit.cloudera.org:8080/10698
Tested-by: Will Berkeley <wdberkeley@gmail.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>