KUDU-2480: tsan failure of master-stress-test
This is an attempt to fix the race reported in KUDU-2480.  I've been
unable to reproduce the race on dist-test, however after auditing the
code mentioned the the stack traces I saw this bit of fishy code. My
current theory is that it's illegal to call
Synchronizer::AsStatusCallback twice on one synchronizer, which was
accidentally happening due to some dead code. This may or may not fix
the race, but it's a mistake worth fixing nonetheless.

Change-Id: I6b5103242aa24ffa81813d8e6ab5df630c69c8cf
Reviewed-on: http://gerrit.cloudera.org:8080/10770
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Dan Burkert <danburkert@apache.org>