tablet: encapsulate common iterator options
I intend to introduce an additional option or two as part of the "diff scan"
API. To that end, this patch knocks the plumbing out of the way by moving
all of the common iterator options into a new struct.

Unfortunately, I can't use it in Tablet::NewRowIterator because while nearly
every iterator expects a pointer to the projection, Tablet::Iterator stores
a copy of the projection itself.

Change-Id: I7232d163436e69999bba75ed66756d2a86c5a959
Reviewed-on: http://gerrit.cloudera.org:8080/10802
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>