[tests] fix flake in one AdminCliTest's scenario
AdminCliTest::TestUnsafeChangeConfigWithMultiplePendingConfigs
had a race condition: leader re-election between calls of
FindTabletLeader() and FindTabletFollowers(), with frequency
about 1/300 if running with dist-test.

The updated code does not contain that race: more then 2K runs
of dist-test passed without a single failure.

In addition to that, this patch contains minor clean-up
related to the RunUnsafeChangeConfig() function.

Change-Id: Iafbaca77d9f0c795bb1787923d50f9fe0a41f334
Reviewed-on: http://gerrit.cloudera.org:8080/10836
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>