KUDU-2492: Make the use of SO_REUSEPORT conditional on it being defined
A recent commit to Kudu, "rpc: add experimental rpc_reuseport flag",
added the use of the rpc flag SO_REUSEREPORT. This flag is not
available with older versions of Linux, resulting in a compiler error.

This patch avoids the compiler error with a macro that checks if
SO_REUSEPORT is defined, and if it's not attempting to set it will
return an error.

Change-Id: I673fa09619955de162e2e2742480672b1e5c4d1d
Reviewed-on: http://gerrit.cloudera.org:8080/10875
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins