[tools] 8 shards for kudu-admin-test
As of now, the kudu-admin-test contains 38 test scenarios.  All of those
can be run independently and in parallel.  If run sequentially, it takes
around 15 minutes to complete those on a powerful server with 48-core CPU.
So, let's bump the number of shards to 8 since many of the scenarios
are long-running ones.

As for the big picture, it would be nice to split the scenarios from
kudu-admin-test and kudu-tools-test to reflect the hierarchy of the
kudu CLI tool itself (i.e. 'cluster', 'diagnose', 'fs', etc.).  Someday.

Change-Id: If1de441749edd9399a690b453fa16c6b0487db45
Reviewed-on: http://gerrit.cloudera.org:8080/10874
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>