loadgen: honor table_num_replicas contract
The description for table_num_replicas says that a value of 0 means use the
server-side default. Except that if you try to use 0 right now, you'll get
an error: "illegal replication factor 0 (replication factor must be odd)".

Passing 0 all the way to the server is rightfully an error; to get the
server-side default, the client needs to leave this field blank.

Change-Id: Ia978b266e4a81a35495e6366f8938ac337c3645c
Reviewed-on: http://gerrit.cloudera.org:8080/10878
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>