consensus: release lock before doing DNS lookups in leader elections
Noticed this while poring over a ~60k thread pstack dump from an overloaded
cluster. Ignoring for a minute that the consensus lock isn't actually needed
to serialize this work, taking a spinlock during a potentially blocking
operation is a no-no. As is doing substantive work in a class constructor;
that's a violation of POLA [1] in our codebase.

1. https://en.wikipedia.org/wiki/Principle_of_least_astonishment

Change-Id: I8403ed74d3199ef90779b86bb7d9727a3c4af876
Reviewed-on: http://gerrit.cloudera.org:8080/10879
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Adar Dembo <adar@cloudera.com>