Define TestKuduMetastorePlugin test data directory at runtime
This is preliminary work needed to run the java tests
on dist-test.

Currently the TestKuduMetastorePlugin test data directory
is defined at build time and passed via a system variable
to the hive-site.xml. This change instead defines the data
directory at test runtime. It also changes the @BeforeClass tag
to a @Before tag to ensure tests are isolated.

Change-Id: I39a81b891b2dd2a7e9d3248facfad4265ec61deb
Reviewed-on: http://gerrit.cloudera.org:8080/10896
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins