tablet: add second snapshot to iterator options
This commit adds a second snapshot parameter called 'snap_to_exclude' and
renames the existing snapshot to 'snap_to_include'. The new snapshot is
hidden behind a boost::optional so that ignoring it is more efficient (i.e.
a call to boost::optional::is_initialized() instead of
MvccSnapshot::IsCommitted() on a snapshot including all transactions).

Taken together, the two snapshots form a time range. If both are set, the
iterator should only show transactions committed in 'snap_to_include' and
not committed in 'snap_to_exclude'. This will be used in the upcoming diff
scan API, which needs to produce results between two HT timestamps.

Change-Id: If4f7390671a637962cc2e3851e8be0e3a6982b17
Reviewed-on: http://gerrit.cloudera.org:8080/10925
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>