memrowset: support iteration with snap_to_exclude
Following on the commit that introduced the snap_to_exclude iterator option,
this commit modifies the MemRowSet iterator to take snap_to_exclude into
consideration (if it is set) when determining which rows and mutations are
relevant to an iterator.

I've also pulled the "exceeded the iterator's scan bounds" short-circuit out
of the MVCC checks. If it were true for a row excluded by MVCC, it'd be true
for all subsequent rows too.

Change-Id: I9be2ab5af4d5223889e4545ae1db7cc0275480ce
Reviewed-on: http://gerrit.cloudera.org:8080/10926
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>