[Java] Fix the use of resource files in tests
This is preliminary work needed to run the java tests
on dist-test.

Our tests currently reference files/resources that are
not available in the output test artifacts. This means
our tests work when they are run from source in the
expected location but fail when they are run from the
compiled artifacts.

This fixes those failing tests by discovering the correct
location via Class.getResource and ensuring all needed
files are packaged in the resources directory.

Change-Id: Ia3a90bac67652b58f2eef802c509983c68d8a299
Reviewed-on: http://gerrit.cloudera.org:8080/10924
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Mike Percy <mpercy@apache.org>