Change yetus dependency to compile in spark modules
Due to Spark’s usage of reflection the Yetus annotation
is required at runtime.

Because it is getting in the way of other development
and Yetus is a small dependency, I changed the scope
to compile in the Spark modules to ensure the class
is available.

See KUDU-2500 for details and to track the
final solution.

Change-Id: I16867a75b7f29f57c557449f94e52439cb9642c9
Reviewed-on: http://gerrit.cloudera.org:8080/10927
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Tony Foerster <anthonymfoerster@gmail.com>