memrowset: support iteration with include_deleted_rows
This is another piece in the incremental backup puzzle. The idea is that
when taking an incremental backup, the scan will include rows that have been
deleted. The backup will figure out which rows were deleted by including a
special "is deleted" virtual column in the projection.

This commit introduces a new iterator option that can be used to include
deleted rows, and uses it in the MemRowSet.

Change-Id: Ie018043518b437ecc719cf9f87b2c5eea560c9a1
Reviewed-on: http://gerrit.cloudera.org:8080/10929
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>