[logging] fix logging init if linking kudu client library
Don't call constructor-like InitializeBasicLogging() function in
non-exported builds of the kudu_client library.  The rationale
is simple: the non-exported kudu_client library is used in Kudu tools,
benchmarks, and tests.  In tools and benchmarks, InitGoogleLoggingSafe()
is called to initialize the logging as needed.  In tests, the gtest
initializes glog on itself.

Prior to this fix, the kudu CLI tool could exit on SIGPIPE when
master/tserver abruptly closed connection.  That's because the call
to InitGoogleLoggingSafe() in the main() function of the CLI tool
didn't do anything once InitGoogleLoggingSafeBasic() has been
inadvertently called by the linked kudu_client library.

Change-Id: I6dcd8d7ccb8a9cafa79674273b95c61eee5baf5a
Reviewed-on: http://gerrit.cloudera.org:8080/10956
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>